% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dockerfile.R
\name{add_dockerfile}
\alias{add_dockerfile}
\title{Add a Dockerfile}
\usage{
add_dockerfile()
}
\value{
a Dockerfile
}
\description{
Adds a Dockerfile to the root directory to containerize a shiny app or report
}
