% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{create a reactable object}
\usage{
get_table(data, size = 10, wrap = FALSE, fullWidth = FALSE, ...)
}
\arguments{
\item{data}{a dataframe}

\item{size}{number of rows to display}

\item{wrap}{wrap columns?}

\item{fullWidth}{should table be full width?}
}
\value{
a reactable
}
\description{
a reactable object
}
