% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_html_format.R
\name{rmd_html_format}
\alias{rmd_html_format}
\title{Custom HTML template}
\usage{
rmd_html_format(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
additional style and template file
}
