% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_numerics.R
\name{round_numerics}
\alias{round_numerics}
\title{Round numeric columns in a dataframe}
\usage{
round_numerics(data, digits = 2)
}
\arguments{
\item{data}{a dataframe}

\item{digits}{number of digits to round to}
}
\value{
a dataframe
}
\description{
Round numeric columns in a dataframe
}
