% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fsa.R
\name{theme_fsa}
\alias{theme_fsa}
\title{apply custom ggplot theme}
\usage{
theme_fsa(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 10,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  font = "IBM Plex Mono",
  ...
)
}
\value{
ggplot theme
}
\description{
fsa ggplot theme
}
